requirements.txt
apache-flink==1.20.2
psycopg2-binary==2.9.1
requests
kafka-python
gtfs-realtime-bindings
numpy
statsmodels